refactor: extract policy module, deduplicate checks, reorder validation
Address review feedback from PR #10314:

- Move policy constants and functions to storage/policy.rs module,
  keeping reject_if_non_standard_tx in storage.rs
- Inline standard_script_kind_from_code (1-line function used once)
  and merge standard_script_kind_from_bytes into standard_script_kind
- Fix doc comments: "zcashd's Solver()" -> "zcash_script solver"
- Reorder checks: scriptSig size/push-only checks now run before
  are_inputs_standard, so count_script_push_ops is never called
  on non-push-only scripts
- Deduplicate spent_outputs length mismatch check: done once in
  reject_if_non_standard_tx instead of in both are_inputs_standard
  and p2sh_sigop_count
- Move unit tests for private helpers into policy module
